Updates to procedure + simulations
1. Added profSim file, used this code to prototype/test the profile
likelihood method.

2. Major updates to the computeMLE function:
a. The `sign’ probabilities can be computed in three ways:
*`all’ - as usual, based on the selected variables and the boundary.
*`selected’ - based only on the selected variables, ignoring the
boundary.
*`onesided’ - condition on the sign and sample only from the selected
sign.

3. Added option to impute the boundaries based on estimated quantities.
a. if neighbors is NULL then the boundary is imputed with the mean of
the estimated selected means.
b. if not NULL, neighbors must be a two-column matrix, the first column
of which indexes the boundary parameters and the second column indexes
their neighbors. The method will perform inference under the assumption
that the mean of the boundary voxels is identical to the mean of the
their neighbors.